string_functions
In coding, strings are sequences of characters located on the same line of code. They're immutable, meaning their value can't be changed without creating a new value altogether. Python comes equipped with a series of string functions that you can use to perform all types of tasks on your website